Added support for calf death loss and dehorning loss.
cull_cows() now honors the involuntary cull rate.

Added a new dependency on the Cerberus library (http://docs.python-cerberus.org/en/stable/).

Added three new parameters. check_all_parms, a Boolean, is used to turn parameter checking on and off. There are now a few dozen of them, and more formal checking is needed to keep them all straight.

calf_loss, a float ranging from 0.0 to 1.0, determines the rate at which calves die during calfhood (the first year of life).

dehorning_loss, a float ranging from 0.0 to 1.0, is used to set the rate at which cows dir during dehorning. The implicit assumption is that all cows are horned and the cow's genotype is not checked. This is useful for certain beef cattle scenarios, but the implementation probably needs to be refined.

culling_rate is only sort-of new. It existed before as an optional parameter to cull_cows(), but was never passed downstream. This is now fixed.

Added a new procedure, check_parameters(), which does parameter checking. The program halts if errors are found, and the errors are printed to stdout.

The involuntary culling term code was changed from "C" to "I". "C" is now used for calfhood deaths.

Added a new procedure, print_death_codes(), in case you can't remember what the codes mean (I sometimes can't).

Stubbed in a new function, disposal_reasons(), which will convert the dead_bulls and dead_cows lists into a Pandas dataframe to do some reporting.