Now only cows with horns will die due to dehorning complications, and only if 'Horned' is one of the conditions in the recessives list.
disposal_reasons() now works, but could be refined to showonly one column of counts, instead of 12 identical columns of counts.

I've started sketching-in code to support nucleus and multiplier herds for beef cattle breeding scenarios. The basic idea is that we'll use two collections of animals -- one small one that is genetically ahead of the larger "commercial" population. Males will move from the nucleus herds to the commercial herds. The gene flow will be only one way, but I think that's a fairly reasonable model for most multiplier herds.